api-harvester/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyGithub==2.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.128.8
uvicorn==0.39.0
pydantic==2.7.0
